refactor: move test files to root-level test/ directory
- Moved all *.test.ts files from src/ to test/
- Updated import paths to use ../../src/
- Added test/AGENTS.md with guidelines for writing CI-reliable tests
- Documents common issues: module mocking, git commands, env vars
